pyproject.toml
[project]
name = "supabase-page-search-mcp"
version = "0.1.0"
description = "MCP server for semantic search of Roam Research pages using OpenAI embeddings and Supabase"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.10.5",
    "httpx>=0.28.1",
    "openai>=1.96.1",
    "python-dotenv>=1.1.1",
]

[project.scripts]
supabase-page-search-mcp = "main:mcp.run"
